face-cut/requirements.txt
boto3 == 1.35.99
pillow
ydb